Module base
Wrote the first base for the complete module. The user is now able to include slidevs and create a slidev. The user is able to set a name, a main layout and a folder with his or her slides. The module is already having a build function and create server function, which are excecuted synchronously and have error catchers. Next up: Building the presentation.