Slidevs address and minor bug fix front-end
The link for the generated Slidev is now coming out of the os of Node, instead of static 'localhost'. Also fixed a minor bug in the front-end, which disallowed the user to slide to the left again after reaching the end of the presentation and still hitted the right arrow button - and the other way around.