Socket drawing, total slides
User is now able to draw on the tablet, where is also appears within the note of the slide. Also build in a check for orientation, so the user can only draw on a note when tablet is in landscape. Also build in a little functionality so the controls know how many slides there are.