Note per slide
Controls are now saving the note per slide. After closing the note of a slide the controls will clean up the canvas in order to add a note to a following slide, except if the note is already drawn. The controls will then redraw the note on the canvas. Also fixed a minor bug which kept on adding new canvases to the slides.