Authentification
User is now able to set a password for the controls. After a valid password check the user is able to keep using the controls for one day. Also wrote the functionality which adds the jquery vendor file to the Slidevs folder, so the controls can also be used when there is not internet connection available. The user is now also able to tap the 'connection' button when red (no connection) to try to reconnect.